# src/api/content-api-openapi.yaml
openapi: 3.0.3
info:
  title: MLG.clan Content Storage and Retrieval API
  description: |
    Comprehensive gaming platform API for content storage, retrieval, and management.
    
    ## Features
    - Content CRUD operations for gaming clips, screenshots, guides, and reviews
    - MLG token integration with Solana blockchain voting system
    - Advanced search and filtering capabilities
    - Content moderation and reporting system
    - Real-time analytics and metrics tracking
    - Scalable file upload/download handling
    - User content management and statistics
    
    ## Authentication
    - JWT tokens for user authentication
    - Solana wallet signature verification for blockchain operations
    - Rate limiting and abuse prevention
    
    ## MLG Token Integration
    - Contract Address: 7XJiwLDrjzxDYdZipnJXzpr1iDTmK55XixSFAa7JgNEL
    - Vote costs: Upvote (1 MLG), Downvote (2 MLG), Super Vote (5 MLG)
    - Reward multipliers based on content type and engagement
    
  version: 1.0.0
  contact:
    name: MLG.clan Development Team
    url: https://mlg.clan/api-support
    email: api-support@mlg.clan
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mlg.clan/v1
    description: Production API Server
  - url: https://staging-api.mlg.clan/v1
    description: Staging API Server
  - url: http://localhost:3000/api/v1
    description: Local Development Server
security:
  - BearerAuth: []
  - WalletSignature: []
paths:
...
  /content:
    get:
      summary: List content with filtering and pagination
      description: |
        Retrieve content list with advanced filtering, sorting, and search capabilities.
        Supports both offset-based and cursor-based pagination for optimal performance.
      tags:
        - Content
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - name: contentType
          in: query
          description: Filter by content type
          schema:
            $ref: '#/components/schemas/ContentType'
        - name: game
          in: query
          description: Filter by game title
          schema:
            type: string
            example: "Fortnite"
        - name: platform
          in: query
          description: Filter by gaming platform
          schema:
            $ref: '#/components/schemas/Platform'
        - name: category
          in: query
          description: Filter by content category
          schema:
            $ref: '#/components/schemas/Category'
        - name: userId
          in: query
          description: Filter by content creator
          schema:
            type: string
            format: uuid
        - name: tags
          in: query
          description: Filter by tags (comma-separated)
          schema:
            type: string
            example: "fortnite,victory-royale,clutch"
        - name: status
          in: query
          description: Filter by content status
          schema:
            $ref: '#/components/schemas/ContentStatus'
        - name: search
          in: query
          description: Full-text search query
          schema:
            type: string
            maxLength: 100
            example: "epic victory royale"
        - name: searchFields
          in: query
          description: Fields to search in
          schema:
            type: string
            enum: [title, description, tags, title,description, title,tags, description,tags, title,description,tags]
            default: "title,description,tags"
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, updated_at, views, likes, trending_score, quality_score]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: createdAfter
          in: query
          description: Filter content created after this date
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          description: Filter content created before this date
          schema:
            type: string
            format: date-time
        - name: minQualityScore
          in: query
          description: Minimum quality score
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: minViews
          in: query
          description: Minimum view count
          schema:
            type: integer
            minimum: 0
        - name: minLikes
          in: query
          description: Minimum like count
          schema:
            type: integer
            minimum: 0
        - name: trending
          in: query
          description: Only return trending content
          schema:
            type: boolean
        - name: featured
          in: query
          description: Only return featured content
          schema:
            type: boolean
      responses:
        '200':
          description: Content list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            type: array
                            items:
                              $ref: '#/components/schemas/Content'
                          pagination:
                            $ref: '#/components/schemas/PaginationResponse'
                          filters:
                            type: object
                            properties:
                              applied:
                                type: object
                                description: Currently applied filters
                              available:
                                type: object
                                description: Available filter options
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
    
    post:
      summary: Create new content
      description: |
        Upload and create new gaming content with metadata and files.
        Supports chunked upload for large files and provides real-time processing status.
      tags:
        - Content
      security:
        - BearerAuth: []
        - WalletSignature: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - metadata
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
                  description: Main content files (max 5 files, 500MB total)
                thumbnail:
                  type: string
                  format: binary
                  description: Optional custom thumbnail image
                metadata:
                  $ref: '#/components/schemas/ContentCreateInput'
            encoding:
              files:
                contentType: video/*, image/*, application/pdf, text/*
              thumbnail:
                contentType: image/*
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            $ref: '#/components/schemas/Content'
                          uploadUrls:
                            type: array
                            items:
                              type: string
                              format: uri
                            description: CDN upload URLs for additional processing
                          processingStatus:
                            type: string
                            enum: [initiated, processing, complete, failed]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /content/{contentId}:
    get:
      summary: Get content by ID
      description: |
        Retrieve specific content with full details, analytics, and related information.
        Optionally increments view count for analytics tracking.
      tags:
        - Content
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/ContentId'
        - name: includeAnalytics
          in: query
          description: Include detailed analytics data
          schema:
            type: boolean
            default: false
        - name: includeComments
          in: query
          description: Include comment preview
          schema:
            type: boolean
            default: false
        - name: includeRelated
          in: query
          description: Include related content suggestions
          schema:
            type: boolean
            default: false
        - name: incrementView
          in: query
          description: Whether to increment view count
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Content retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            $ref: '#/components/schemas/Content'
                          analytics:
                            $ref: '#/components/schemas/ContentAnalytics'
                          comments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Comment'
                          relatedContent:
                            type: array
                            items:
                              $ref: '#/components/schemas/Content'
                          userVote:
                            $ref: '#/components/schemas/Vote'
                          downloadUrls:
                            type: object
                            additionalProperties:
                              type: string
                              format: uri
                            description: Temporary download URLs for content files
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
    
    put:
      summary: Update content
      description: |
        Update content metadata, settings, and file associations.
        Only content owners can perform updates.
      tags:
        - Content
      security:
        - BearerAuth: []
        - WalletSignature: []
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentUpdateInput'
      responses:
        '200':
          description: Content updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          content:
                            $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete content
      description: |
        Permanently delete content and associated files.
        Only content owners or administrators can delete content.
      tags:
        - Content
      security:
        - BearerAuth: []
        - WalletSignature: []
      parameters:
        - $ref: '#/components/parameters/ContentId'
        - name: reason
          in: query
          description: Deletion reason for analytics
          schema:
            type: string
            maxLength: 200
        - name: hardDelete
          in: query
          description: Permanently delete vs soft delete
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Content deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Content Voting System
  /content/{contentId}/vote:
    post:
      summary: Vote on content with MLG tokens
      description: |
        Cast vote by burning MLG tokens with Solana blockchain verification.
        Requires valid wallet signature and sufficient token balance.
      tags:
        - Voting
      security:
        - BearerAuth: []
        - WalletSignature: []
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoteInput'
      responses:
        '201':
          description: Vote cast successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          vote:
                            $ref: '#/components/schemas/Vote'
                          contentStats:
                            type: object
                            properties:
                              totalVotes:
                                type: integer
                              newTrendingScore:
                                type: number
                                format: float
                          userStats:
                            type: object
                            properties:
                              tokensRemaining:
                                type: number
                                format: float
                              totalVotesGiven:
                                type: integer
        '400':
          description: Invalid vote data or insufficient tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_tokens:
                  summary: Insufficient MLG tokens
                  value:
                    success: false
                    error: "VOTE_001"
                    message: "Insufficient MLG tokens"
                    details:
                      required: 1
                      available: 0.5
                invalid_signature:
                  summary: Invalid transaction signature
                  value:
                    success: false
                    error: "VOTE_003"
                    message: "Invalid transaction signature"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Already voted on this content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "VOTE_002"
                message: "Already voted on this content"
                details:
                  existingVoteId: "vote-123e4567-e89b-12d3-a456-426614174000"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /content/trending:
    get:
      summary: Get trending content
      description: |
        Retrieve trending content based on engagement metrics and MLG token activity.
        Algorithm considers views, votes, comments, and recent activity.
      tags:
        - Discovery
      security:
        - BearerAuth: []
        - {}
      parameters:
        - name: timeframe
          in: query
          description: Trending timeframe
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
        - name: game
          in: query
          description: Filter by specific game
          schema:
            type: string
        - name: category
          in: query
          description: Filter by content category
          schema:
            $ref: '#/components/schemas/Category'
        - name: limit
          in: query
          description: Number of trending items to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Trending content retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          trending:
                            type: array
                            items:
                              $ref: '#/components/schemas/Content'
                          trendingMetrics:
                            type: object
                            properties:
                              timeframe:
                                type: string
                              totalEngagement:
                                type: integer
                              totalTokenActivity:
                                type: number
                                format: float

  # Content Moderation
  /content/{contentId}/report:
    post:
      summary: Report content for moderation
      description: |
        Submit content report for community moderation.
        Reports are reviewed by moderators and may result in content removal.
      tags:
        - Moderation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportInput'
      responses:
        '201':
          description: Report submitted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          reportId:
                            type: string
                            format: uuid
                          status:
                            type: string
                            enum: [pending, reviewing]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Report rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "RATE_003"
                message: "Report limit exceeded"
                details:
                  resetTime: "2024-01-01T13:00:00Z"

  # User Content Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet-Signature
      description: |
        Solana wallet signature for blockchain operations.
        Required format: {signature}:{message}:{publicKey}
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    Cursor:
      name: cursor
      in: query
      description: Cursor for pagination
      schema:
        type: string
  schemas:
...
    ContentType:
      type: string
      enum:
        - video_clip
        - screenshot
        - guide
        - review
        - stream
        - tournament_content
      example: video_clip
...